# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "tech-software-engineer-consultant-int-backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/SI/add-SI": {
      "post": {
        "tags": [
          "SI"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SICreateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
...
    "/api/SI/get-SI-by-id/{id}": {
      "get": {
        "tags": [
          "SI"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SIDTO"
            }
          }
        }
      }
    },
    "/api/SI/update-SI/{SIId}": {
      "put": {
        "tags": [
          "SI"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "SIId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SIUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/api/SI/delete-SI/{id}": {
      "delete": {
        "tags": [
          "SI"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
    "/api/Transaction/addTransaction": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ReferenceCommande",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TransactionCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BooleanInt32ValueTuple"
            }
          }
        }
      }
    },
...
    "/api/Transaction/getTransactionbyRef/{ReferenceCommande}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "ReferenceCommande",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TransactionsDTO"
              }
            }
          }
        }
      }
    },
    "/api/Transaction/getTransactionbyId/{id}": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TransactionsDTO"
            }
          }
        }
      }
    },
    "/api/Transaction/updateTransaction/{id}": {
      "put": {
        "tags": [
          "Transaction"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "TransactionId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TransactionsUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BooleanInt32ValueTuple"
            }
          }
        }
      }
    },
    "/api/Transaction/deleteTransactionbyId/{id}": {
      "delete": {
        "tags": [
          "Transaction"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "boolean"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BooleanInt32ValueTuple": {
      "type": "object",
      "additionalProperties": false
    },
...
    "SICreateDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "adresse_Image_SI": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "SIDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "adresse_Image_SI": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "SIUpdateDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "adresse_Image_SI": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "TransactionCreateDto": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "nomProduit": {
          "type": "string"
        },
        "prixUnitaire": {
          "format": "double",
          "type": "number"
        },
        "quantityEntered": {
          "format": "int32",
          "type": "integer"
        },
        "ptht": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "pttc": {
          "format": "double",
          "type": "number"
        },
        "typeTransaction": {
          "$ref": "#/definitions/TypeTransaction"
        }
      },
      "additionalProperties": false
    },
...
    "TransactionsDTO": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "referenceCommande": {
          "type": "string"
        },
        "referenceLot": {
          "type": "string"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "nomProduit": {
          "type": "string"
        },
        "prixUnitaire": {
          "format": "double",
          "type": "number"
        },
        "quantityEntered": {
          "format": "int32",
          "type": "integer"
        },
        "ptht": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "pttc": {
          "format": "double",
          "type": "number"
        },
        "typeTransaction": {
          "$ref": "#/definitions/TypeTransaction"
        },
        "remainingQuantity": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "TransactionsUpdateDTO": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "nomProduit": {
          "type": "string"
        },
        "prixUnitaire": {
          "format": "double",
          "type": "number"
        },
        "quantityEntered": {
          "format": "int32",
          "type": "integer"
        },
        "ptht": {
          "format": "double",
          "type": "number"
        },
        "tva": {
          "format": "double",
          "type": "number"
        },
        "pttc": {
          "format": "double",
          "type": "number"
        },
        "typeTransaction": {
          "$ref": "#/definitions/TypeTransaction"
        }
      },
      "additionalProperties": false
    },
    "TypeTransaction": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
...
  }
}